src/test/java/br/edu/iff/sistemaacademico/domain/usecase/implementation/GetSubjectUseCaseImplTest.java
package br.edu.iff.sistemaacademico.domain.usecase.implementation;

import br.edu.iff.sistemaacademico.domain.dto.RequestSubject;
import br.edu.iff.sistemaacademico.domain.entity.Subject;
import br.edu.iff.sistemaacademico.domain.usecase.CreateSubjectUseCase;
import br.edu.iff.sistemaacademico.repository.SubjectRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;

import java.util.UUID;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static org.mockito.internal.verification.VerificationModeFactory.times;

@SpringBootTest(classes = GetSubjectUseCaseImpl.class)
@EnableAutoConfiguration(exclude = {DataSourceAutoConfiguration.class, HibernateJpaAutoConfiguration.class})
public class GetSubjectUseCaseImplTest {

    @MockBean
    private SubjectRepository repository;

    @Autowired
    private GetSubjectUseCaseImpl getSubjectUseCase;

    @Test
    public void testPerform() {
        // Given
        String subjectId = "123";
        Subject subject = new Subject();
        subject.setId(subjectId);

        when(repository.findById(subjectId)).thenReturn(subject);
        // When
        Subject result = getSubjectUseCase.perform(subjectId);

        // Then
        assertEquals(subject, result);
        verify(repository, times(1)).findById(subjectId);
    }
}
